refactor calls toolbaar to own component
